- Added shown statistics for smoked joints and spent money - Added the "Add" activity along with the Smoke a joint fragment [WIP]

+++ app/src/main/java/com/gami/highcompanion/data/DaysDAO.kt
@@ -24,11 +24,11 @@ interface DaysDAO {
 
 
     @Query("SELECT spent_amount FROM days_table where sql_date = DATE('now')")
-    fun getDailySpentAmount(): LiveData<Float>
+    fun getDailySpentAmount(): LiveData<Double>
 
     @Query("SELECT SUM(spent_amount) FROM days_table WHERE sql_date >= DATE('now', 'weekday 0', '-7 days')  AND sql_date != DATE('now')  AND sql_date != DATE('now','-1 day')")
-    fun getWeeklySpentAmount(): LiveData<Float>
+    fun getWeeklySpentAmount(): LiveData<Double>
 
     @Query("SELECT SUM(spent_amount) FROM days_table where strftime('%W',sql_date) != strftime('%W',date('now')) AND strftime('%Y',sql_date) = strftime('%Y',date('now')) AND  strftime('%m',sql_date) = strftime('%m',date('now')) AND sql_date != DATE('now', 'weekday 0', '-7 days')  AND sql_date != DATE('now')  AND sql_date != DATE('now','-1 day')")
-    fun getMonthlySpentAmount(): LiveData<Float>
+    fun getMonthlySpentAmount(): LiveData<Double>
 }
